TRUEFIBACKEND/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
openai==1.2.4
httpx==0.25.2
python-dotenv==1.0.0
pyjwt==2.8.0
pydantic==2.5.0
bcrypt==4.1.2
requests==2.31.0 